+++ src/options/components/Account/SignOutModal/index.ts
@@ -1,0 +1,1 @@
+export { SignOutModal } from './SignOutModal';